+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/thumbnail/Thumbnail.kt
@@ -1,4 +1,4 @@
-package org.depromeet.clog.server.domain.thumbnail.domain
+package org.depromeet.clog.server.domain.thumbnail
 
 data class Thumbnail(
     val id: Long? = null,
