+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/thumbnail/ThumbnailRepository.kt
@@ -1,0 +1,5 @@
+package org.depromeet.clog.server.domain.thumbnail
+
+interface ThumbnailRepository {
+    fun save(thumbnail: Thumbnail): Thumbnail
+}
