--- clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/thumbnail/infrastructure/thumbnail/ThumbnailRepository.kt
@@ -1,7 +1,0 @@
-package org.depromeet.clog.server.domain.thumbnail.infrastructure.thumbnail
-
-import org.depromeet.clog.server.domain.thumbnail.domain.Thumbnail
-
-interface ThumbnailRepository {
-    fun save(thumbnail: Thumbnail): Thumbnail
-}
