+++ clog-infrastructure/src/main/kotlin/org/depromeet/clog/server/infrastructure/thumbnail/ThumbnailServiceImpl.kt
@@ -2,11 +2,10 @@ package org.depromeet.clog.server.infrastructure.thumbnail
 
 import org.depromeet.clog.server.domain.thumbnail.application.ThumbnailService
 import org.depromeet.clog.server.domain.thumbnail.application.ThumbnailUploadResponse
-import org.depromeet.clog.server.domain.thumbnail.domain.Thumbnail
-import org.depromeet.clog.server.domain.thumbnail.infrastructure.thumbnail.ThumbnailRepository
+import org.depromeet.clog.server.domain.thumbnail.Thumbnail
+import org.depromeet.clog.server.domain.thumbnail.ThumbnailRepository
 import org.depromeet.clog.server.domain.video.VideoRepository
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
 import org.springframework.web.multipart.MultipartFile
 
 @Service
